requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
asgiref==3.4.1
    # via django
django==3.2.24
    # via djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
pytz==2021.3
    # via
    #   django
    #   djangorestframework
sqlparse==0.4.4
    # via django
